requirements.txt
behave==1.2.6
selenium==4.15.2
requests==2.31.0
pytest==7.4.3
allure-behave==2.13.2
pandas==2.1.3
webdriver-manager==4.0.1
PyHamcrest==2.0.4
pyotp==2.9.0
cryptography==41.0.5
python-dotenv==1.0.0
faker==20.1.0
